cluster 1: main files

+++ Path_Planning/ClassNode.h
@@ -13,7 +13,7 @@ class Node {		// Node = state in Koeing
 public:
 	//char Name; //debug
 	int X, Y;
-	uint8_t cost;   //read from height map (from which we will derive edge costs)
+	int cost;   //read from height map (from which we will derive edge costs)
 	float g, rhs;	//g = cost function,	rhs = one step lookahead value of g
 	//std::vector<float> g, rhs;
 	NodeTypes nodeType;
@@ -21,13 +21,13 @@ public:
 	std::pair<float, float> key;
 
 	std::vector<Sptr_toNode> AdjacentsVect;	//all nodes adjacent to current one (const??)
-	robin_hood::unordered_map < Sptr_toNode, uint8_t > parents;    //key: ptr to node, value: cumulative cost
-	//robin_hood::unordered_map < Sptr_toNode, std::vector<uint8_t> > parents;    //key: ptr to node, value: cumulative cost(s)
+	robin_hood::unordered_map < Sptr_toNode, int > parents;    //key: ptr to node, value: cumulative cost
+	//robin_hood::unordered_map < Sptr_toNode, std::vector<int> > parents;    //key: ptr to node, value: cumulative cost(s)
 
 ////////////////////////////////////   Constructors   /////////////////////////////////
 	Node() {}		// for pointers etc.
 
-	Node(int x, int y, uint8_t ec, NodeTypes flag)  // for actual nodes
+	Node(int x, int y, int ec, NodeTypes flag)  // for actual nodes
 		: X(x), Y(y), cost(ec), nodeType(flag) {
 
 		g = std::numeric_limits<float>::infinity();
@@ -56,12 +56,6 @@ public:
 				return false;
 			else { // k1==k1 & k2==k2	->	to allow different nodes with the same key to be both in the queue
 				//if (X != N2.X || Y != N2.Y)  -> CRASH -> comparator has to be strict
-
-				//if (nodeType == start)// to avoid premature termination in computeMOPaths() -> if Start has same key of top-node, execute! (so appears >)
-				//	return true;			// for a strict weak ordering, comp(x, x) must be false !!!
-				//else if (N2.nodeType == start)
-				//	return false;
-
 				if (X != N2.X)			
 					return (X > N2.X);	//arbitrary order (not important)
 				else if (Y != N2.Y)
@@ -105,12 +99,12 @@ class dummyNode {
 public:
 	//char Name; //debug
 	int X, Y;
-	uint8_t cost;
+	int cost;
 	NodeTypes nodeType;
 
 /////////////////////////   Constructors   //////////////////////////
 	dummyNode() {}		// for pointers etc.
 
-	dummyNode(int x, int y, uint8_t ec, NodeTypes flag) // for actual nodes
+	dummyNode(int x, int y, int ec, NodeTypes flag) // for actual nodes
 		: X(x), Y(y), cost(ec), nodeType(flag) {}
 };
